backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.28.0
gunicorn==21.2.0
flask-socketio==5.3.6
eventlet==0.35.2
pandas==2.2.0
openpyxl==3.1.2
